requirements.txt
streamlit
google-generativeai
transformers
torch
plotly
python-dotenv
deep-translator
textblob
altair
requests
pandas
streamlit-mic-recorder